plot mseed data
